pyproject.toml
[tool.poetry]
name = "fakebank"
version = "0.0.1"
description = ""
authors = ["Yakov Varnaev <yakov.varnaev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
websockets = "^12.0"
uvicorn = "^0.27.1"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.3"}
asyncpg = "^0.29.0"
pydantic-settings = "^2.1.0"
alembic = "^1.13.1"
aiokafka = "^0.10.0"
aioredis = "^2.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
mypy = "^1.8.0"
axblack = "^20220330"
httpx = "^0.27.0"
pytest = "^8.0.1"
polyfactory = "^2.14.1"
anyio = "^4.3.0"

[tool.black]
line-length = 80

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
